# apps/api/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ConfirmEmailRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  dto.ErrorResponse:
    properties:
      detail:
        description: set when a body failed binding or validation.
        type: string
      error:
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
      remember_me:
        type: boolean
    required:
    - email
    - password
    type: object
  dto.LogoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  dto.RateLimitResponse:
    properties:
      error:
        type: string
      retry_after_seconds:
        type: integer
    type: object
  dto.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.ResendVerificationRequest:
    properties:
      email:
        maxLength: 255
        type: string
    required:
    - email
    type: object
...
  dto.TokenResponse:
    properties:
      access_expires_at:
        type: string
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/dto.UserBrief'
    type: object
...
  dto.UserBrief:
    properties:
      account_id:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
...
info:
  contact: {}
  description: Every /v1 route outside /public needs an access token. The active branch
    travels in the X-Branch-Id header, not in the token, and is validated per request.
  title: Coti API
  version: "1.0"
paths:
...
  /v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Advances the session epoch and revokes the presented refresh family.
        The body is optional.
      parameters:
      - description: Refresh token to revoke
        in: body
        name: body
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Logged out
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log out
      tags:
      - auth
...
  /v1/public/auth/login:
    post:
      consumes:
      - application/json
      description: A bad email, a bad password, a disabled user and a deactivated
        account all answer 401 alike. The refresh token is shown exactly once.
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: A confirmed address is required and this one is not
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Rate limit spent; retry_after_seconds says when
          schema:
            $ref: '#/definitions/dto.RateLimitResponse'
      summary: Log in
      tags:
      - auth
  /v1/public/auth/refresh:
    post:
      consumes:
      - application/json
      description: Replaying a consumed token inside the grace window is a benign
        race; past it, the whole family is revoked as theft.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Rotate a refresh token
      tags:
      - auth
  /v1/public/auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Always answers 202, so the response reveals nothing about which
        addresses exist or which are already confirmed. Requesting a new link retires
        the previous one.
      parameters:
      - description: Address to send the link to
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Request a new confirmation link
      tags:
      - auth
...
  /v1/public/auth/verify-email:
    post:
      consumes:
      - application/json
      description: The link works once and expires. Unknown, expired and already-used
        tokens all answer 401 alike. Confirming an address that is already verified
        succeeds.
      parameters:
      - description: Verification token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ConfirmEmailRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Address confirmed
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Confirm an email address
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Access token, prefixed with "Bearer ".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"